Crate sqlx_sqlite

SQLite database driver.

Note: linkage is semver-exempt.

This driver uses the libsqlite3-sys crate which links the native library for SQLite 3. With the “sqlite” feature, we enable the bundled feature which builds and links SQLite from source.

We reserve the right to upgrade the version of libsqlite3-sys as necessary to pick up new 3.x.y versions of SQLite.

Due to Cargo’s requirement that only one version of a crate that links a given native library exists in the dependency graph at a time, using SQLx alongside another crate linking libsqlite3-sys like rusqlite is a semver hazard.

If you are doing so, we recommend pinning the version of both SQLx and the other crate you’re using to prevent a cargo update from breaking things, e.g.:

sqlx = { version = "=0.8.1", features = ["sqlite"] }
rusqlite = "=0.32.1"

and then upgrade these crates in lockstep when necessary.

Dynamic linking

To dynamically link to a system SQLite library, the “sqlite-unbundled” feature can be used instead.

This allows updating SQLite independently of SQLx or using forked versions, but you must have SQLite installed on the system or provide a path to the library at build time (See the rusqlite README for details).

It may result in link errors if the SQLite version is too old. Version 3.20.0 or newer is recommended. It can increase build time due to the use of bindgen.

Modules

any

types

Conversions between Rust and SQLite types.

Structs

LockedSqliteHandle

Sqlite

Sqlite database driver.

SqliteArguments

SqliteColumn

SqliteConnectOptions

Options and flags which can be used to configure a SQLite connection.

SqliteConnection

A connection to an open Sqlite database.

SqliteError

SqliteOwnedBuf

Memory buffer owned and allocated by SQLite. Freed on drop.

SqliteQueryResult

SqliteRow

Implementation of Row for SQLite.

SqliteStatement

SqliteTransactionManager

Implementation of [TransactionManager] for SQLite.

SqliteTypeInfo

Type information for a SQLite type.

SqliteValue

SqliteValueRef

UpdateHookResult

Enums

SqliteArgumentValue

SqliteAutoVacuum

SqliteJournalMode

Refer to SQLite documentation for the meaning of the database journaling mode.

SqliteLockingMode

Refer to SQLite documentation for the meaning of the connection locking mode.

SqliteOperation

SqliteSynchronous

Refer to SQLite documentation for the meaning of various synchronous settings.

Traits

SqliteExecutor

An alias for Executor<'_, Database = Sqlite>.

Type Aliases

SqlitePool

An alias for Pool, specialized for SQLite.

SqlitePoolOptions

An alias for PoolOptions, specialized for SQLite.

SqliteTransaction

An alias for Transaction, specialized for SQLite.